requirements.txt
scrapegraphai==1.11.3
langchain-google-genai==1.0.8
google-generativeai==0.7.2
playwright==1.45.1
python-dotenv==1.0.1
